+++ docs/release_notes.rst
@@ -8,6 +8,7 @@ Release notes for **Shopit**.
 0.1.1
 =====
 
+* Make fields in product serializer editable through settings, set optimized defaults.
 * Fix error when mergin dictionaries in python3.
 * Remove redundant code.
 * Fix trying to generate image thumbnail on attachment when `file` is None.